Lab 8/requirements.txt
pandas
scikit-learn
boto3
apache-airflow